Add files via upload
Thank you very much for your recognition of our work. As the technology used in this work is currently being applied in a Chinese company and related technologies are being applied for national invention patents, the complete data and code may affect the smooth development of future work. Therefore, the complete code cannot be released at this time, and we apologize for any inconvenience this may cause.
However, in consideration of the needs of other researchers, we have provided approximately 80% of the source code used in this paper in the supplementary documentation, with the remaining 20% described in detail in the document using annotation. Readers can replicate our model more easily by referring to the existing code and the detailed description in the paper.
Besides, We left the contact information in supplementary documentation and explained the reasons for retaining some of the code. If researchers have academic or teaching needs, we welcome them to send us an email. We will evaluate whether the reader's use is only for academic or teaching purposes before choosing to share all the code separately. 
E-mail: mengkunphd@163.com